kaReducer was moved tto root path
