+++ src/Demos/AlertCellDemo/AlertCellDemo.tsx
@@ -2,9 +2,8 @@ import './AlertCellDemo.scss';
 
 import React, { useState } from 'react';
 
-import { ITableProps, Table } from '../../lib';
+import { ITableProps, kaReducer, Table } from '../../lib';
 import { DataType } from '../../lib/enums';
-import { kaReducer } from '../../lib/reducers';
 import { CellFuncPropsWithChildren, DispatchFunc } from '../../lib/types';
 
 const dataArray = Array(10).fill(undefined).map(
