+++ src/Demos/CustomDataRowDemo/CustomDataRowDemo.tsx
@@ -1,8 +1,7 @@
 import React, { useState } from 'react';
 
-import { ITableProps, Table } from '../../lib';
+import { ITableProps, kaReducer, Table } from '../../lib';
 import { DataType, SortDirection, SortingMode } from '../../lib/enums';
-import { kaReducer } from '../../lib/reducers';
 import { DataRowFuncPropsWithChildren, DispatchFunc } from '../../lib/types';
 
 const dataArray: any[] = [
