+++ src/Demos/DeleteRowDemo/DeleteRowDemo.tsx
@@ -2,10 +2,9 @@ import './DeleteRowDemo.scss';
 
 import React, { useState } from 'react';
 
-import { ITableProps, Table } from '../../lib';
+import { ITableProps, kaReducer, Table } from '../../lib';
 import { deleteRow } from '../../lib/actionCreators';
 import { DataType } from '../../lib/enums';
-import { kaReducer } from '../../lib/reducers';
 import { CellFuncPropsWithChildren, DispatchFunc } from '../../lib/types';
 
 const dataArray = Array(10).fill(undefined).map(
