+++ src/Demos/Demos.scss
@@ -49,7 +49,7 @@ body {
     }
 
     .resources-icons {
-      padding: 20px 0 0 40px;
+      padding: 40px 0 0 40px;
       height: 28px;
       border-top: 1px solid #CBDCE4;
       display: flex;
