+++ src/Demos/EventsDemo/EventsDemo.tsx
@@ -2,10 +2,9 @@ import './EventsDemo.scss';
 
 import React, { useState } from 'react';
 
-import { ITableProps, Table } from '../../lib';
+import { ITableProps, kaReducer, Table } from '../../lib';
 import { DataType, EditingMode, FilteringMode, SortingMode } from '../../lib/enums';
 import { ChildAttributes } from '../../lib/models';
-import { kaReducer } from '../../lib/reducers';
 import { DispatchFunc } from '../../lib/types';
 import { EventsLog } from './EventsLog';
