+++ src/Demos/SearchDemo/SearchDemo.tsx
@@ -1,9 +1,8 @@
 import React, { useState } from 'react';
 
-import { ITableProps, Table } from '../../lib';
+import { ITableProps, kaReducer, Table } from '../../lib';
 import { search } from '../../lib/actionCreators';
 import { DataType } from '../../lib/enums';
-import { kaReducer } from '../../lib/reducers';
 import { DispatchFunc } from '../../lib/types';
 
 const dataArray: any[] = [
