+++ src/Demos/SelectionSingleDemo/SelectionSingleDemo.tsx
@@ -2,11 +2,10 @@ import './SelectionSingleDemo.scss';
 
 import React, { useState } from 'react';
 
-import { ITableProps, Table } from '../../lib';
+import { ITableProps, kaReducer, Table } from '../../lib';
 import { deselectAllRows, selectSingleRow } from '../../lib/actionCreators';
 import { ActionType, DataType } from '../../lib/enums';
 import { ChildAttributes } from '../../lib/models';
-import { kaReducer } from '../../lib/reducers';
 import { DispatchFunc } from '../../lib/types';
 import dataArray from './data';
