+++ src/lib/index.ts
@@ -1,1 +1,2 @@
 export * from './Components/Table/Table';
+export * from './Reducers/kaReducer';
