--- src/lib/reducers.ts
@@ -1,1 +1,0 @@
-export * from './Reducers/kaReducer';
